trying to cast

+++ HW05/task2.cu
@@ -17,8 +17,8 @@ int main() {
     cudaMemset(dA, 0, n * sizeof(int));
 
     mt19937 generator(12354);
-    uniform_real_distribution<int> dist(0, 100);
-    auto r = dist(generator);
+    uniform_real_distribution<float> dist(0., 100.);
+    auto r = static_cast <int> dist(generator);
 
     algebraKernel<<<2, 8>>>(dA, r);
     cudaDeviceSynchronize();
